@@ -14,7 +14,7 @@
     }
   ],
   "require": {
-    "php": ">=7.3",
+    "php": ">=7.4",
     "illuminate/support": "5.0.*|5.1.*|5.2.*|5.3.*|5.4.*|5.5.*|5.6.*|5.7.*|5.8.*|^6.0|^7.0|^8.0|^9.0",
     "illuminate/validation": "5.0.*|5.1.*|5.2.*|5.3.*|5.4.*|5.5.*|5.6.*|5.7.*|5.8.*|^6.0|^7.0|^8.0|^9.0",
     "illuminate/translation": "5.0.*|5.1.*|5.2.*|5.3.*|5.4.*|5.5.*|5.6.*|5.7.*|5.8.*|^6.0|^7.0|^8.0|^9.0"
@@ -22,7 +22,8 @@
   "require-dev": {
     "mockery/mockery": "^1.4.2",
     "phpunit/phpunit": "^9.0",
-    "symfony/var-dumper": "5.1.5.*@dev"
+    "symfony/var-dumper": "5.1.5.*@dev",
+    "imanghafoori/eloquent-mockery": "dev-main"
   },
   "autoload": {
     "psr-4": {
@@ -36,7 +37,10 @@
     "psr-4": {
       "Tests\\": "tests/",
       "Tests\\Helpers\\": "tests/helpers/"
-    }
+    },
+    "files": [
+      "tests/helpers/functions.php"
+    ]
   },
   "minimum-stability": "dev",
   "extra": {
