@@ -1,28 +1,16 @@
-<?php use Illuminate\Contracts\Events\Dispatcher;
+<?php
 
-/**
- * Created by PhpStorm.
- * User: darryl
- * Date: 1/12/2015
- * Time: 9:59 PM
- */
-
+use Illuminate\Contracts\Events\Dispatcher;
 use Mohamadtsn\ShoppingCart\Cart;
-use Mockery as m;
 use Tests\Helpers\SessionMock;
-
 
 class CartTestOtherFormat extends PHPUnit\Framework\TestCase
 {
-
-    /**
-     * @var Mohamadtsn\ShoppingCart\Cart
-     */
-    protected $cart;
+    protected Cart $cart;
 
     public function setUp(): void
     {
-        $events = m::mock(Dispatcher::class);
+        $events = Mockery::mock(Dispatcher::class);
         $events->shouldReceive('dispatch');
 
         $this->cart = new Cart(
@@ -36,7 +24,7 @@ class CartTestOtherFormat extends PHPUnit\Framework\TestCase
 
     public function tearDown(): void
     {
-        m::close();
+        Mockery::close();
     }
 
     public function test_cart_sub_total()
